Warn user of coins created in previous wallet version
